+++ src/channel_reporter/reporter.rs
@@ -2,11 +2,21 @@ use crate::{EventReporter, EventSender};
 use std::error;
 use std::fmt::{Debug, Display, Formatter};
 
+/// Proof that a [`ChannelReporter`] has been disconnected.
+///
+/// Produced by [`ChannelReporter::disconnect`] and consumed by
+/// [`ChannelHandlerGuard::join`]. Because the inner field is private
+/// and this type has no public constructor, the only way to obtain a `DisconnectToken` value
+/// is by calling [`ChannelReporter::disconnect`], which should guarantee correct ordering.
+///
+/// [`ChannelHandlerGuard::join`]: crate::ChannelHandlerGuard::join
+pub struct DisconnectToken(());
+
 /// A specialized type of reporter which uses a channel to transmit messages.
 ///
 /// Use [`EventReporter::disconnect`] to disconnect the channel by dropping the `sender`.
-/// If you want to finish up processing unprocessed events, you may do so by calling
-/// the blocking [`FinishProcessing::finish_processing`].
+/// This returns a [`DisconnectToken`] proof token which must be passed to
+/// [`HandlerGuard::join`] to wait for the handler thread to drain and exit.
 ///
 /// The channel sender (and channel receiver for the `listener`), required to create a
 /// `ChannelReporter` instance can be created by calling the [`event_channel()`]
@@ -15,7 +25,7 @@ use std::fmt::{Debug, Display, Formatter};
 /// The [`EventListener`] associated with this reporter is the [`ChannelEventListener`].
 ///
 /// [`EventReporter::disconnect`]: crate::EventReporter::disconnect
-/// [`FinishProcessing::finish_processing`]: crate::FinishProcessing::finish_processing
+/// [`HandlerGuard::join`]: crate::HandlerGuard::join
 /// [`event_channel()`]: crate::event_channel()
 /// [`EventListener`]: crate::EventListener
 /// [`ChannelEventListener`]: crate::ChannelEventListener
@@ -36,6 +46,7 @@ impl<Event> ChannelReporter<Event> {
 impl<Event> EventReporter for ChannelReporter<Event> {
     type Event = Event;
     type Err = EventReporterError<Event>;
+    type DisconnectToken = DisconnectToken;
 
     fn report_event(&self, event: impl Into<Self::Event>) -> Result<(), Self::Err> {
         self.event_sender
@@ -43,10 +54,15 @@ impl<Event> EventReporter for ChannelReporter<Event> {
             .map_err(EventReporterError::SendError)
     }
 
-    /// Disconnect the sender.
-    #[allow(clippy::unit_arg)]
-    fn disconnect(self) -> Result<(), Self::Err> {
-        Ok(self.event_sender.disconnect())
+    /// Disconnect the sender, returning a [`DisconnectToken`] token.
+    ///
+    /// Pass the token to [`HandlerGuard::join`] to wait for the handler thread to finish
+    /// draining the queue. The token enforces that this call happens first.
+    ///
+    /// [`HandlerGuard::join`]: crate::HandlerGuard::join
+    fn disconnect(self) -> Result<DisconnectToken, Self::Err> {
+        self.event_sender.disconnect();
+        Ok(DisconnectToken(()))
     }
 }